# Refactor trailer Field for Improved Memory Efficiency and Performance